Refactored Slash commands into their own class making interactions easier to read
